requirements.txt
openai
beautifulsoup4
requests
lxml
snowflake-connector-python
streamlit
python-dotenv
